Defined legality checking predicates and refactored legalizeMoves().
Defined the predicates mentioned in the last commit and re-wrote legalizeMoves() using lambda expressions to deal with the loop index related bug. Deleting elements of the vector in the for() loop would have required extra control logic due to the changing size of the vector throughout the duration of the loop. This method is simpler and far easier to maintain.